# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ELSA API",
        "version": "1.0"
    },
    "paths": {
        "/v1/auth/register/email": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Регистрация по электронной почте",
                "description": "Регистрация пользователя по электронной почте.",
                "operationId": "b754180a5f634a91e34d4e7e0c22b2ee",
                "requestBody": {
                    "description": "Данные пользователя для регистрации",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secretpassword"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная регистрация"
                    },
                    "400": {
                        "description": "Некорректный запрос"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера"
                    }
                }
            }
        }
    }
}